-removed the cmm_err messages (nfs4_pnfs.c) -fixed a recursive mutex locking issue (nfs4_pnfs.c) -fixed a potential NULL pointer dereference (pnfs_get_device in nfs4_pnfs.c) -fixed statistics accounting when client has multiple mounts from the same server (nfs4_vfsops.c) -moved the check for pNFS from pnfs_read/write to nfs4read/nfs4write functions (nfs4_vnops.c)

+++ usr/src/cmd/fs.d/nfs/nfsstat/nfsstat_layout.h
@@ -114,7 +114,7 @@ extern bool_t xdr_stripe_info_t(XDR *xdrs, stripe_info_t *objp);
 extern bool_t xdr_layoutstats_t(XDR *xdrs, layoutstats_t *objp);
 static int create_client_handle(char *address, struct netconfig *nconf,
     ulong_t prog, ulong_t vers, CLIENT **client);
-extern int lookup_name_port(netaddr4 *na, int *port,
+extern int lookup_name_port(netaddr4 *na, long *port,
     char *hostname, char *ipaddress);
 extern int null_procedure_ping(char *netid, char *address,
     enum clnt_stat *ds_status);
